pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "routingtools"
version = "0.1.0"
description = "Excalidraw-as-config: parse routing topology diagrams into runtime config (YAML/JSON/CSV/Mermaid)."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["excalidraw", "routing", "config-generation", "diagram-as-code"]
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
routingtools = "routingtools.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
